HH-70 <projectDetails and projectDetailsTransport CRUD and testing in Postman>
