--- src/main/java/com/hamdihawari/server/language/controller/LanguageController.java
@@ -1,39 +1,0 @@
-package com.hamdihawari.server.language.controller;
-
-import com.hamdihawari.server.language.entity.Language;
-import com.hamdihawari.server.language.service.LanguageService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/language")
-public class LanguageController {
-
-    @Autowired
-    private LanguageService languageService;
-
-    @GetMapping
-    public List<Language> getAllLanguages() {
-        return languageService.getAllLanguages();
-    }
-
-    @GetMapping("/{id}")
-    public ResponseEntity<Language> getLanguageById(@PathVariable Long id) {
-        Language language = languageService.getLanguageById(id);
-        return language != null ? ResponseEntity.ok(language) : ResponseEntity.notFound().build();
-    }
-
-    @PostMapping
-    public Language createLanguage(@RequestBody Language language) {
-        return languageService.saveLanguage(language);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteLanguage(@PathVariable Long id) {
-        languageService.deleteLanguage(id);
-        return ResponseEntity.noContent().build();
-    }
-}
