--- src/main/java/com/hamdihawari/server/language/entity/Language.java
@@ -1,33 +1,0 @@
-
-
-package com.hamdihawari.server.language.entity;
-
-import jakarta.persistence.*;
-
-@Entity
-@Table(name = "language")
-public class Language {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String name; // Name der Sprache, z.B. "English", "Deutsch", "عربي"
-
-    // Getter und Setter
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
