+++ src/main/java/com/hamdihawari/server/project/language/dto/LanguageDTO.java
@@ -1,0 +1,25 @@
+
+package com.hamdihawari.server.project.language.dto;
+
+public class LanguageDTO {
+
+    private Long id;
+    private String name;
+
+    // Getter und Setter
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
