+++ src/main/java/com/hamdihawari/server/project/language/entity/Language.java
@@ -1,0 +1,39 @@
+package com.hamdihawari.server.project.language.entity;
+
+import jakarta.persistence.*;
+
+@Entity
+@Table(name = "language")
+public class Language {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String name;
+
+    public Language() {
+    }
+
+    // Constructor for creating Language with ID
+    public Language(Long id) {
+        this.id = id;
+    }
+
+    // Getters and Setters...
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
