+++ src/main/java/com/hamdihawari/server/project/language/repository/LanguageRepository.java
@@ -1,0 +1,7 @@
+package com.hamdihawari.server.project.language.repository;
+
+import com.hamdihawari.server.project.language.entity.Language;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface LanguageRepository extends JpaRepository<Language, Long> {
+}
