+++ src/main/java/com/hamdihawari/server/project/language/service/LanguageService.java
@@ -1,0 +1,37 @@
+
+package com.hamdihawari.server.project.language.service;
+
+import com.hamdihawari.server.project.language.entity.Language;
+import com.hamdihawari.server.project.language.repository.LanguageRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class LanguageService {
+
+    @Autowired
+    private LanguageRepository languageRepository;
+
+    public List<Language> getAllLanguages() {
+        return languageRepository.findAll();
+    }
+
+    public Language getLanguageById(Long id) {
+        return languageRepository.findById(id).orElse(null);
+    }
+
+    public Language saveLanguage(Language language) {
+        return languageRepository.save(language);
+    }
+
+    public void deleteLanguage(Long id) {
+        languageRepository.deleteById(id);
+    }
+
+    public Language findById(Long id) {
+        return languageRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Language not found with id: " + id));
+    }
+}
