+++ src/main/java/com/hamdihawari/server/project/projectCard/repository/ProjectCardTranslationRepository.java
@@ -1,0 +1,6 @@
+package com.hamdihawari.server.project.projectCard.repository;
+
+import com.hamdihawari.server.project.projectCard.entity.ProjectCardTranslation;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProjectCardTranslationRepository extends JpaRepository<ProjectCardTranslation, Long> {}
