--- src/main/java/com/hamdihawari/server/projectCard/repository/ProjectCardRepository.java
@@ -1,9 +1,0 @@
-package com.hamdihawari.server.projectCard.repository;
-
-import com.hamdihawari.server.projectCard.entity.ProjectCard;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ProjectCardRepository extends JpaRepository<ProjectCard, Long> {
-}
