--- src/main/java/com/hamdihawari/server/projectCard/repository/ProjectCardTranslationRepository.java
@@ -1,6 +1,0 @@
-package com.hamdihawari.server.projectCard.repository;
-
-import com.hamdihawari.server.projectCard.entity.ProjectCardTranslation;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProjectCardTranslationRepository extends JpaRepository<ProjectCardTranslation, Long> {}
